+++ src/main/java/kitchenpos/orders/application/OrderService.java
@@ -1,15 +1,15 @@
-package kitchenpos.application;
+package kitchenpos.orders.application;
 
 import kitchenpos.menus.domain.model.Menu;
 import kitchenpos.menus.domain.model.MenuRepository;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
-import kitchenpos.domain.OrderType;
-import kitchenpos.infra.KitchenridersClient;
+import kitchenpos.orders.domain.model.Order;
+import kitchenpos.orders.domain.model.OrderLineItem;
+import kitchenpos.orders.domain.model.OrderRepository;
+import kitchenpos.orders.domain.model.OrderStatus;
+import kitchenpos.orders.domain.model.OrderTable;
+import kitchenpos.orders.domain.model.OrderTableRepository;
+import kitchenpos.orders.domain.model.OrderType;
+import kitchenpos.orders.infrastructure.KitchenridersClient;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
