+++ src/main/java/kitchenpos/orders/domain/model/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.orders.domain.model;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
