+++ src/main/java/kitchenpos/orders/infrastructure/JpaOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.orders.infrastructure;
 
+import kitchenpos.orders.domain.model.Order;
+import kitchenpos.orders.domain.model.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
