+++ src/main/java/kitchenpos/orders/infrastructure/JpaOrderTableRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.orders.infrastructure;
 
+import kitchenpos.orders.domain.model.OrderTableRepository;
+import kitchenpos.orders.domain.model.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
