+++ src/test/java/kitchenpos/application/FakeKitchenridersClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.infra.KitchenridersClient;
+import kitchenpos.orders.infrastructure.KitchenridersClient;
 
 import java.math.BigDecimal;
 import java.util.UUID;
