+++ src/test/java/kitchenpos/application/InMemoryOrderRepository.java
@@ -1,9 +1,9 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.orders.domain.model.Order;
+import kitchenpos.orders.domain.model.OrderRepository;
+import kitchenpos.orders.domain.model.OrderStatus;
+import kitchenpos.orders.domain.model.OrderTable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
