+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,13 +1,14 @@
 package kitchenpos.application;
 
 import kitchenpos.menus.domain.model.MenuRepository;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
-import kitchenpos.domain.OrderType;
+import kitchenpos.orders.application.OrderService;
+import kitchenpos.orders.domain.model.Order;
+import kitchenpos.orders.domain.model.OrderLineItem;
+import kitchenpos.orders.domain.model.OrderRepository;
+import kitchenpos.orders.domain.model.OrderStatus;
+import kitchenpos.orders.domain.model.OrderTable;
+import kitchenpos.orders.domain.model.OrderTableRepository;
+import kitchenpos.orders.domain.model.OrderType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
